ESWE-67: Move current work activity logic to JPA entity object
